// generated/src/app/api-models/geo-region-zip-range.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class GeoRegionZipRange {
  seqGeoRegionZipRange: number;
  regionType: string;
  regionCode: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}